changes in create table to also update the secondary indexes catalog
