+++ docs/journal/part_8_v2_random_notes.md
